use stubs classes in pytypes test
